LayoutTests/webrtc/video-unmute-expected.txt



PASS Setting video exchange
PASS Track is disabled, video should be black
PASS Track is enabled, video should not be black
